+++ test.sh
@@ -14,8 +14,10 @@ mkdir -p output
 source ./_setup.sh
 
 # Execute tests
-docker-compose exec -T test env \
-    CONJUR_AUTHN_API_KEY="$api_key" \
-    CONJUR_V4_AUTHN_API_KEY="$api_key_v4" \
-    CONJUR_V4_SSL_CERTIFICATE="$ssl_cert_v4" \
-    bash -c 'go test -v $(go list ./... | grep -v /vendor/) | tee output/junit.output && cat output/junit.output | go-junit-report > output/junit.xml'
+CONJUR_AUTHN_API_KEY="$api_key" \
+  CONJUR_V4_AUTHN_API_KEY="$api_key_v4" \
+  CONJUR_V4_SSL_CERTIFICATE="$ssl_cert_v4" \
+  docker-compose run test \
+    bash -o pipefail \
+      -c 'go test -v $(go list ./... | grep -v /vendor/) | tee output/junit.output && cat output/junit.output | go-junit-report > output/junit.xml' || echo "TESTS FAILED"
+